+++ src/font/unicode-bidi-tables.h
@@ -42,7 +42,7 @@
 
 typedef struct _bidichar_type_map
 {
-    Uint32    type;       // Type of Unicode characters
+    Uint16    type;       // Type of Unicode characters
     Uchar32   chv;        // Starting index of Unicode character
     Uint8     count;      // Total number of Unicode characters of same type starting from glyphn
 } BIDICHAR_TYPE_MAP;
